shell32.go: change the wrapper for Shell_NotifyIcon to obtain the last error code from syscall
As outlined in https://github.com/tailscale/tailscale/issues/4133#issuecomment-1320351481,
the syscall wrappers in this package should be obtaining the last error value
directly from syscall, but are not.

This patch fixes the one for Shell_NotifyIcon in an effort to help diagnose its
failures.

Updates https://github.com/tailscale/tailscale/issues/4133

Signed-off-by: Aaron Klotz <aaron@tailscale.com>