shell32.go: change Shell_NotifyIcon definition back to returning bool instead of error
Even though Windows docs do not make any mention of this API returning an error
code, we tried anyway for the purposes of diagnosing another bug. It never
really helped, and eventually we did fix that bug (the API was being called
off the UI goroutine). Let's revert the definition back to match its
documentation so that we don't get thrown off by potentially bogus error codes.

This is #cleanup.

Signed-off-by: Aaron Klotz <aaron@tailscale.com>